apps/api/requirements.txt
fastapi
openai
qdrant-client
pydantic
python-dotenv
uvicorn
langchain
pytest
httpx
google-generativeai
cohere
